# did/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gov-DID Judicial Collaboration Backend API
  version: 1.0.0
  description: 第五章六个后端模块的核心 REST 契约；支撑接口以实际 Router 为准。
servers:
  - url: http://localhost:8081
security:
  - bearerAuth: []
paths:
  /api/system/session:
    get:
      summary: 获取登录账号、管理员权限与当前链上执行主体
      parameters: [{$ref: '#/components/parameters/ActorAlias'}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
...
  /api/did/query:
    get:
      summary: 查询 DID 注册状态
      parameters: [{$ref: '#/components/parameters/DID'}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
...
  /api/privacy/group:
    get:
      summary: 查询当前群组锚与成员公钥
      parameters: [{$ref: '#/components/parameters/GroupID'}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
...
  /api/revocation/logs:
    get:
      summary: 查询凭证撤销日志和最终状态
      parameters: [{$ref: '#/components/parameters/VCID'}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  parameters:
    ActorAlias: {name: X-Actor-Alias, in: header, required: false, description: 仅管理员可指定的已配置链上执行主体, schema: {type: string}}
    DID: {name: did, in: query, required: true, schema: {type: string}}
    GroupID: {name: groupID, in: query, required: true, schema: {type: string}}
    VCID: {name: vcID, in: query, required: true, schema: {type: string}}
  responses:
    Success:
      description: 统一成功响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
  schemas:
    Envelope:
      type: object
      required: [code, message, data, requestId]
      properties:
        code: {type: string, example: OK}
        message: {type: string, example: success}
        data: {type: object, additionalProperties: true}
        requestId: {type: string}
        tx:
          type: object
          properties:
            txId: {type: string}
            blockHeight: {type: integer, format: int64}
            gasUsed: {type: integer, format: int64}